backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
joblib==1.5.1
librosa==0.11.0
numpy==2.1.3
scipy==1.16.1
tensorflow-cpu==2.19.0
lightgbm==4.6.0
gunicorn
soundfile==0.13.1